+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.10.6
 
+* `Versions`
+    * `Ktor`: `2.0.1` -> `2.0.2`
 * `Common`
     * `JS`:
         * Add `ResizeObserver` functionality